src/RotateKeyFunc/rotate_key/__init__.py
import datetime
import logging

import azure.functions as func


def main(rotation_timer: func.TimerRequest) -> None:
    utc_timestamp = datetime.datetime.utcnow().replace(
        tzinfo=datetime.timezone.utc).isoformat()

    if rotation_timer.past_due:
        logging.info('The timer is past due!')

    logging.info('Python timer trigger function ran at %s', utc_timestamp)
