src/RotateKeyFunc/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
atomicwrites==1.4.0
attrs==20.3.0
azure-core==1.10.0
azure-identity==1.5.0
azure-storage-blob==12.7.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
flake8==3.8.4
flake8-docstrings==1.5.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
mccabe==0.6.1
msal==1.8.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.0
packaging==20.8
pluggy==0.13.1
portalocker==1.7.1
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
urllib3==1.26.3
